Make CI Pass (#11)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Makes some edits to our code to make sure that our unit tests and
linters pass, so that we can conveniently use CI instead of having to
look for new and unusual problems in it.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Having CI is actually amazing
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: XeonMations <62395746+XeonMations@users.noreply.github.com>
Co-authored-by: FalloutFalcon <falloutfalconplays@gmail.com>