Implement aggravated damage (#28)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Modern TG has done away with CLONE damage since they no longer have
cloning as a feature. This is a problem because the old WoD codebase
made heavy use of CLONE damage as a stand-in for aggravated damage as it
appears in tabletop. This PR implements a new AGGRAVATED damage type for
supernatural damage so that we can give it special handling, damage
mechanics, and healing mechanics.

This adds full compatibility for dealing and healing aggravated damage,
along with adding it to relevant places like examine and a component we
use for Disciplines.

This does not add aggravated damage dealing and healing (or even
reading, in the case of medical scanners) to every place it'd make
sense. I'd rather not bloat the PR more, I'll just leave that for when
those other bits of code get used.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
If we don't want to have our aggravated damage type be BURN, this is a
major project we need to undergo as part of the rebase.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
refactor: Aggravated damage no longer uses clone damage, and instead has
its own unique damage type.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->