Stat Wrappers and storyteller stat normalization (#58)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

- Normalizes the discrepancy between storyteller stats being called
traits interchangeably, by making us use 'stats' for the name primarily.

- Creates wrapper functions for quickly calling storyteller stats on
living mobs

- Makes all functions for editing storyteller stats begin with the
variable referencing the stat, followed by the amount if applicable,
followed by additional parameters.

- Breaks out the defines for the ability and attribute datums into their
own file, so its children can be deactivated on a by-file basis in
downstreams

- Adds a base_type to all attributes and abilities, so stat datums do
not load the attribute and ability base type.

- Adds previously undefined pool attribute types; no functionality set
for them as of yet.

- Stats that alter max health alter the player's current health
dynamically, giving them temporary health bonuses in the case of modded
values.

## Why It's Good For The Game

The our attribute code is going to be used very often in our code. It
makes sense that we should streamline it and improve readability on the
code.

Also, due to the attributes themselves likely diverging between code
bases, we should make it easier to add new ones without modifying the
base code overmuch.

---------

Co-authored-by: FalloutFalcon <86381784+FalloutFalcon@users.noreply.github.com>